+++ src/pure-function-helpers/xml-to-json.ts
@@ -1,6 +1,6 @@
 import { parseString } from 'xml2js'
 
-export async function xml2json (xml: string): Promise<any> {
+export async function xmlToJson (xml: string): Promise<any> {
   return new Promise((resolve, reject) => {
     parseString(xml, { explicitArray: false }, (err, result) => {
       if (err) {
